With icons ( not thumnails )
Document Directory component 
==>In Tutorials Section - Displays Video Icons
Once we click on any of the video, The video will play in the video column section. ( like youtube video playing in a 3/4 screen )
and remaining videos should be available beside in the playlist column. ( Displays only Icons with Title )
==>In Document Section - Displays File icon with their FileType and Title.
Once we click on any of the file, It will download automatically.